spa_auth_app/requirements.txt
asgiref==3.3.1
astroid==2.5.1
boto3==1.17.32
botocore==1.20.32
Django==3.1.7
django-cors-headers==3.7.0
django-static-md5url==0.1
django-storages==1.11.1
djangorestframework==3.12.2
isort==5.7.0
jmespath==0.10.0
lazy-object-proxy==1.5.2
mccabe==0.6.1
psycopg2-binary==2.8.6
pylint==2.7.2
python-dateutil==2.8.1
pytz==2021.1
s3transfer==0.3.6
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.4
uWSGI==2.0.19.1
wrapt==1.12.1
